admin pages

+++ database/migrations/2023_07_22_143549_create_cards_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->string("name");
             $table->string("image");
-            $table->float("price");
+            $table->decimal('price', 10, 2);
             $table->foreignIdFor(category::class)->constrained()->onDelete('cascade');;
             $table->boolean("avilability");
             $table->boolean('require_id');
